Refactor Digilent HS3 program
