+++ CHANGELOG.rst
@@ -1,6 +1,10 @@
 avr32-prog
 ==========
 
+**2.0.0 (not released)**
+
+- Feature: Refactor digilent hs3 documentation
+
 **1.0.0 (2023-09-07)**
 
 - Feature: First release of fork
