+++ digilent_hs3/requirements.txt
@@ -1,0 +1,2 @@
+pyelftools>=0.29
+pyftdi>=0.54.0
